test: align repository test layout
