+++ src/repo.rs
@@ -173,64 +173,8 @@ mod tests {
 
         assert_eq!(get_current_branch_name(&repo).unwrap(), "detached HEAD");
     }
-
-    #[test]
-    fn test_clone_error_reports_missing_requested_branch() {
-        let missing_reference = git2::Error::new(
-            ErrorCode::NotFound,
-            ErrorClass::Reference,
-            "reference not found",
-        );
-        assert_eq!(
-            clone_error_message(
-                "owner/repo",
-                Some("missing"),
-                &missing_reference
-            ),
-            "The specified branch 'missing' does not exist in the repository."
-        );
-        assert_eq!(
-            clone_error_message("owner/repo", None, &missing_reference),
-            "Failed to clone: reference not found; class=Reference (4); code=NotFound (-3)"
-        );
-    }
-
-    #[test]
-    fn test_clone_error_reports_network_context() {
-        let network = git2::Error::new(
-            ErrorCode::GenericError,
-            ErrorClass::Net,
-            "offline",
-        );
-        assert_eq!(
-            clone_error_message("owner/repo", None, &network),
-            "Network error: The repository 'owner/repo' might not exist or you may not have permission to access it."
-        );
-    }
-
-    #[test]
-    fn test_clone_error_reports_authentication_guidance() {
-        let authentication = git2::Error::new(
-            ErrorCode::Auth,
-            ErrorClass::Http,
-            "too many redirects or authentication replays",
-        );
-        assert_eq!(
-            clone_error_message("owner/private", None, &authentication),
-            "The repository 'owner/private' does not exist or requires authentication.\nIf it's a private repository, please provide a valid token using the --token option."
-        );
-    }
-
-    #[test]
-    fn test_clone_error_preserves_unexpected_details() {
-        let unexpected = git2::Error::new(
-            ErrorCode::GenericError,
-            ErrorClass::Os,
-            "disk full",
-        );
-        assert_eq!(
-            clone_error_message("owner/repo", None, &unexpected),
-            "Failed to clone: disk full; class=Os (2)"
-        );
-    }
 }
+
+#[cfg(test)]
+#[path = "../tests/crate/repo.rs"]
+mod clone_error_tests;
